Move CreateObject validation into plugin
