+++ src/Scope.ts
@@ -17,16 +17,7 @@ import { URI } from 'vscode-uri';
 import { LogLevel } from './Logger';
 import type { BrsFile } from './files/BrsFile';
 import type { DependencyGraph, DependencyChangedEvent } from './DependencyGraph';
-import { isBrsFile, isClassMethodStatement, isClassStatement, isCustomType, isEnumStatement, isFunctionStatement, isFunctionType, isLiteralExpression, isXmlFile } from './astUtils/reflection';
-import { nodes, components } from './roku-types';
-import type { BRSComponentData } from './roku-types';
-import type { Token } from './lexer/Token';
-
-/**
- * The lower-case names of all platform-included scenegraph nodes
- */
-const platformNodeNames = new Set(Object.values(nodes).map(x => x.name.toLowerCase()));
-const platformComponentNames = new Set(Object.values(components).map(x => x.name.toLowerCase()));
+import { isBrsFile, isClassMethodStatement, isClassStatement, isCustomType, isEnumStatement, isFunctionStatement, isFunctionType, isXmlFile } from './astUtils/reflection';
 
 /**
  * A class to keep track of all declarations within a given scope (like source scope, component scope)
@@ -240,7 +231,7 @@ export class Scope {
      * Get the list of files referenced by this scope that are actually loaded in the program.
      * Includes files from this scope and all ancestor scopes
      */
-    public getAllFiles() {
+    public getAllFiles(): BscFile[] {
         return this.cache.getOrAdd('getAllFiles', () => {
             let result = [] as BscFile[];
             let dependencies = this.dependencyGraph.getAllDependencies(this.dependencyGraphKey);
@@ -503,76 +494,7 @@ export class Scope {
             this.diagnosticDetectFunctionCollisions(file);
             this.detectVariableNamespaceCollisions(file);
             this.diagnosticDetectInvalidFunctionExpressionTypes(file);
-            this.validateCreateObjectCalls(file);
         });
-    }
-
-    /**
-     * Validate every function call to `CreateObject`.
-     * Ideally we would create better type checking/handling for this, but in the mean time, we know exactly
-     * what these calls are supposed to look like, and this is a very common thing for brs devs to do, so just
-     * do this manually for now.
-     */
-    protected validateCreateObjectCalls(file: BrsFile) {
-        for (const call of file.functionCalls) {
-            if (call.name?.toLowerCase() === 'createobject' && isLiteralExpression(call?.args[0]?.expression)) {
-                const firstParamToken = (call?.args[0]?.expression as any)?.token;
-                const firstParamStringValue = firstParamToken?.text?.replace(/"/g, '');
-                //if this is a `createObject('roSGNode'` call, only support known sg node types
-                if (firstParamStringValue?.toLowerCase() === 'rosgnode' && isLiteralExpression(call?.args[1]?.expression)) {
-                    const componentName: Token = (call?.args[1]?.expression as any)?.token;
-                    //add diagnostic for unknown components
-                    const unquotedComponentName = componentName?.text?.replace(/"/g, '');
-                    if (unquotedComponentName && !platformNodeNames.has(unquotedComponentName.toLowerCase()) && !this.program.getComponent(unquotedComponentName)) {
-                        this.diagnostics.push({
-                            file: file as BscFile,
-                            ...DiagnosticMessages.unknownRoSGNode(unquotedComponentName),
-                            range: componentName.range
-                        });
-                    } else if (call?.args.length !== 2) {
-                        // roSgNode should only ever have 2 args in `createObject`
-                        this.diagnostics.push({
-                            file: file as BscFile,
-                            ...DiagnosticMessages.mismatchCreateObjectArgumentCount(firstParamStringValue, [2], call?.args.length),
-                            range: call.range
-                        });
-                    }
-                } else if (!platformComponentNames.has(firstParamStringValue.toLowerCase())) {
-                    this.diagnostics.push({
-                        file: file as BscFile,
-                        ...DiagnosticMessages.unknownBrightScriptComponent(firstParamStringValue),
-                        range: firstParamToken.range
-                    });
-                } else {
-                    // This is valid brightscript component
-                    // Test for invalid arg counts
-                    const brightScriptComponent: BRSComponentData = components[firstParamStringValue.toLowerCase()];
-                    // Valid arg counts for createObject are 1+ number of args for constructor
-                    let validArgCounts = brightScriptComponent.constructors.map(cnstr => cnstr.params.length + 1);
-                    if (validArgCounts.length === 0) {
-                        // no constructors for this component, so createObject only takes 1 arg
-                        validArgCounts = [1];
-                    }
-                    if (!validArgCounts.includes(call?.args.length)) {
-                        // Incorrect number of arguments included in `createObject()`
-                        this.diagnostics.push({
-                            file: file as BscFile,
-                            ...DiagnosticMessages.mismatchCreateObjectArgumentCount(firstParamStringValue, validArgCounts, call?.args.length),
-                            range: call.range
-                        });
-                    }
-
-                    // Test for deprecation
-                    if (brightScriptComponent.isDeprecated) {
-                        this.diagnostics.push({
-                            file: file as BscFile,
-                            ...DiagnosticMessages.deprecatedBrightScriptComponent(firstParamStringValue, brightScriptComponent.deprecatedDescription),
-                            range: call.range
-                        });
-                    }
-                }
-            }
-        }
     }
 
     /**
