+++ src/bscPlugin/validation/ScopeValidator.ts
@@ -1,13 +1,22 @@
 import { Location } from 'vscode-languageserver';
 import { URI } from 'vscode-uri';
-import { isBrsFile } from '../../astUtils/reflection';
+import { isBrsFile, isLiteralExpression } from '../../astUtils/reflection';
 import { Cache } from '../../Cache';
 import { DiagnosticMessages } from '../../DiagnosticMessages';
 import type { BrsFile } from '../../files/BrsFile';
-import type { BsDiagnostic, OnScopeValidateEvent } from '../../interfaces';
+import type { BscFile, BsDiagnostic, OnScopeValidateEvent } from '../../interfaces';
 import type { DottedGetExpression } from '../../parser/Expression';
 import type { EnumStatement } from '../../parser/Statement';
 import util from '../../util';
+import { nodes, components } from '../../roku-types';
+import type { BRSComponentData } from '../../roku-types';
+import type { Token } from '../../lexer/Token';
+
+/**
+ * The lower-case names of all platform-included scenegraph nodes
+ */
+const platformNodeNames = new Set(Object.values(nodes).map(x => x.name.toLowerCase()));
+const platformComponentNames = new Set(Object.values(components).map(x => x.name.toLowerCase()));
 
 export class ScopeValidator {
     constructor(
@@ -18,8 +27,8 @@ export class ScopeValidator {
     public process() {
         this.validateEnumUsage();
         this.detectDuplicateEnums();
+        this.validateCreateObjectCalls();
     }
-
 
     /**
      * Find all expressions and validate the ones that look like enums
@@ -114,6 +123,80 @@ export class ScopeValidator {
                 });
             }
         }
+        this.event.scope.addDiagnostics(diagnostics);
+    }
+
+    /**
+     * Validate every function call to `CreateObject`.
+     * Ideally we would create better type checking/handling for this, but in the mean time, we know exactly
+     * what these calls are supposed to look like, and this is a very common thing for brs devs to do, so just
+     * do this manually for now.
+     */
+    protected validateCreateObjectCalls() {
+        const diagnostics = [];
+        this.event.scope.enumerateBrsFiles((file) => {
+            for (const call of file.functionCalls) {
+                //skip non CreateObject function calls
+                if (call.name?.toLowerCase() !== 'createobject' || !isLiteralExpression(call?.args[0]?.expression)) {
+                    continue;
+                }
+                const firstParamToken = (call?.args[0]?.expression as any)?.token;
+                const firstParamStringValue = firstParamToken?.text?.replace(/"/g, '');
+                //if this is a `createObject('roSGNode'` call, only support known sg node types
+                if (firstParamStringValue?.toLowerCase() === 'rosgnode' && isLiteralExpression(call?.args[1]?.expression)) {
+                    const componentName: Token = (call?.args[1]?.expression as any)?.token;
+                    //add diagnostic for unknown components
+                    const unquotedComponentName = componentName?.text?.replace(/"/g, '');
+                    if (unquotedComponentName && !platformNodeNames.has(unquotedComponentName.toLowerCase()) && !this.event.program.getComponent(unquotedComponentName)) {
+                        diagnostics.push({
+                            file: file as BscFile,
+                            ...DiagnosticMessages.unknownRoSGNode(unquotedComponentName),
+                            range: componentName.range
+                        });
+                    } else if (call?.args.length !== 2) {
+                        // roSgNode should only ever have 2 args in `createObject`
+                        diagnostics.push({
+                            file: file as BscFile,
+                            ...DiagnosticMessages.mismatchCreateObjectArgumentCount(firstParamStringValue, [2], call?.args.length),
+                            range: call.range
+                        });
+                    }
+                } else if (!platformComponentNames.has(firstParamStringValue.toLowerCase())) {
+                    diagnostics.push({
+                        file: file as BscFile,
+                        ...DiagnosticMessages.unknownBrightScriptComponent(firstParamStringValue),
+                        range: firstParamToken.range
+                    });
+                } else {
+                    // This is valid brightscript component
+                    // Test for invalid arg counts
+                    const brightScriptComponent: BRSComponentData = components[firstParamStringValue.toLowerCase()];
+                    // Valid arg counts for createObject are 1+ number of args for constructor
+                    let validArgCounts = brightScriptComponent.constructors.map(cnstr => cnstr.params.length + 1);
+                    if (validArgCounts.length === 0) {
+                        // no constructors for this component, so createObject only takes 1 arg
+                        validArgCounts = [1];
+                    }
+                    if (!validArgCounts.includes(call?.args.length)) {
+                        // Incorrect number of arguments included in `createObject()`
+                        diagnostics.push({
+                            file: file as BscFile,
+                            ...DiagnosticMessages.mismatchCreateObjectArgumentCount(firstParamStringValue, validArgCounts, call?.args.length),
+                            range: call.range
+                        });
+                    }
+
+                    // Test for deprecation
+                    if (brightScriptComponent.isDeprecated) {
+                        diagnostics.push({
+                            file: file as BscFile,
+                            ...DiagnosticMessages.deprecatedBrightScriptComponent(firstParamStringValue, brightScriptComponent.deprecatedDescription),
+                            range: call.range
+                        });
+                    }
+                }
+            }
+        });
         this.event.scope.addDiagnostics(diagnostics);
     }
 }
